Extract HTTP client adapter tests
